refactor: implement comprehensive code quality improvements for lawyer-chat
- Break down monolithic LawyerChatContent component (1097 lines) into modular components
  - Created MessageList component for message rendering
  - Created ChatInput component for user input handling
  - Created ChatControls component for download functionality
  - Created useChatState hook for centralized state management
  - Created useChatAPI hook for all API interactions

- Implement type safety improvements
  - Added runtime validation with Zod schemas
  - Created type guards for error handling
  - Replaced type assertions with validated parsing
  - Fixed all 'any' type usage

- Centralize configuration and constants
  - Created src/config/constants.ts for all magic numbers
  - Fixed scattered configuration across files
  - Proper naming for all constants with units

- Implement safe coding practices
  - Added Redis-based rate limiting with fallback
  - Fixed memory leak from unused response.json()
  - Added SSR guards for window access
  - Proper input validation on all API endpoints
  - Comprehensive error handling with type guards

All existing functionality preserved and tested.
