+++ services/lawyer-chat/package.json
@@ -65,6 +65,7 @@
     "rehype-raw": "^7.0.0",
     "remark-gfm": "^4.0.1",
     "remark-math": "^6.0.0",
+    "zod": "^4.0.5",
     "zustand": "^5.0.5"
   },
   "devDependencies": {
