+++ services/lawyer-chat/src/types/chat.ts
@@ -1,0 +1,10 @@
+import type { AnalyticsData } from '@/types';
+
+export interface Message {
+  id: number;
+  sender: 'user' | 'assistant';
+  text: string;
+  references?: string[];
+  analytics?: AnalyticsData;
+  timestamp: Date;
+}
